+++ ReduxArea.ts
@@ -1,10 +1,10 @@
-import produce from "immer"
+import produce, { Draft, Immutable, Produced } from "immer"
 import { AnyAction, Reducer } from 'redux'
 
 type Func = (...args: any) => any
 type ReduxAction = ((...args: any) => AnyAction) & { name: string; reducer: Reducer }
 
-const CreateReduxArea = <TState>(initialState: TState) => {
+const CreateReduxArea = <TState, D = Draft<TState>>(initialState: TState) => {
    const actions: ReduxAction[] = []
    return {
       add: (name: string) => ({
@@ -14,29 +14,29 @@ const CreateReduxArea = <TState>(initialState: TState) => {
                ...action.apply(null, args),
                type: name
             })
-            const mappedAction = actionCreator as unknown as ((...args: Parameters<typeof action>) => ActionType) & {
+            const mappedAction = actionCreator as unknown as ((...args: Parameters<typeof action>) => ReturnType<T> & { type: string }) & {
                name: string,
-               reducer: Reducer<TState, ActionType>
-               type: ActionType
+               reducer: Reducer<Immutable<TState>, ReturnType<T> & { type: string }>
+               type: ReturnType<T> & { type: string }
             }
             Object.defineProperty(mappedAction, 'name', {
                value: name,
                writable: false
             })
             return {
-               produce: (producer: (draft: TState, action: ActionType) => void) => {
+               produce: (producer: (draft: Draft<TState>, action: ReturnType<T> & { type: string }) => void) => {
                   Object.defineProperty(mappedAction, 'reducer', {
-                     value: produce(producer) as Reducer<TState, ActionType>,
+                     value: produce(producer) as Reducer<Immutable<TState>, ReturnType<T> & { type: string }>,
                      writable: false
                   })
                   actions.push(mappedAction as unknown as ReduxAction)
                   return mappedAction;
                },
                reducer: (
-                  reducer: (state: TState, reducerAction: ActionType) => any | void
+                  reducer: (state: TState, reducerAction: ReturnType<T> & { type: string }) => any | void
                ) => {
                   Object.defineProperty(mappedAction, 'reducer', {
-                     value: reducer as Reducer<TState, ActionType>,
+                     value: reducer as Reducer<TState, ReturnType<T> & { type: string }>,
                      writable: false
                   })
                   actions.push(mappedAction as unknown as ReduxAction)