+++ CHANGES.md
@@ -1,4 +1,11 @@
-## 0.0.7
+## 0.0.9
+
+Fix problem with readonly typescript and immer produce.
+
+The draft will now ignore readonly and the produce will return the correct typescript difinition
+which is a immutable state. (Immer make the next state immutable)
+
+## 0.0.7-0.0.8
 
 Add action.type for type definition to use in custom Saga's Reducer ect.